Added Stat Component and Settings + Misc additions
The stat component is a node you can add in the world to house specific
stats. These stats are stored as resources for easier serialization.

This system is designed to be modular, allowing for things like health,
stamina, mana, currency, anything you want to track can be done with
this system. In the demo scene I have added a two areas that add and
subtract from health which is being shown on screen with a ProgressBar
node

Crouch speed was missing and has been implemented